publish/pyproject.toml
[project]
name = "publish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=5.5.0",
    "duckdb==1.2.1",
    "google-cloud-bigquery>=3.31.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pyiceberg>=0.9.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
repl = [
    "ipykernel>=6.29.5",
]
